requirements.txt
asgiref==3.5.2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
docker==7.0.0
dotenv-cli==3.3.0
executing==2.0.1
idna==3.7
iniconfig==2.0.0
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.4
pluggy==1.5.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pymongo==3.12.1
pytest==8.2.2
pytest-django==4.8.0
pytest-testdox==3.1.0
python-dotenv==1.0.1
pytz==2022.1
pywin32==306
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
